update users bug.
This is how it should be working:

Similar way to how the chat updates. I want the users to be placed in a
text file and be updated every second.. That part works, but its
somewhere in the chat.js and process.php that the error is occuring.

Im trying to append 2 different text files. but for some reason its
appending the first one to the second box, while deleting the first
one. Im guessing theres some sort of issue with the